PIMS-1876 NumberFormatException: For input string: "c879370c3b259eba013b620ec0cd0352"

+++ grails-app/taglib/org/pih/warehouse/CategorySelectTagLib.groovy
@@ -67,7 +67,7 @@ class CategorySelectTagLib {
 	}
 	
 	def categorySelect = { attrs ->		
-		def selectedCategory = Category.get(attrs.value as int)
+		def selectedCategory = Category.get(attrs.value)
 		def rootCategory = productService.getRootCategory();
 		def excludeSpaces = attrs?.excludeSpaces
